TAG <section> must have tag <h1...h6>
While it is useful to provide a heading for each sectioning content element, it is not required.

The only way to prevent this warning when using sectioning content elements is, well, to follow the advice by providing a heading content element. If you’d use a div instead, your document outline would be different and likely wrong (if you used section correctly in the first place).